"Add methods to attach, remove, and update roles in AuthLiteClient class; add error checking for role ID types"
This commit adds new methods to the AuthLiteClient class for attaching, removing, and updating roles. It also includes error checking to ensure that the role IDs are of the correct type. The methods make API requests to the AuthLite server to assign or remove roles for a user. The 'signoff\_session\_and\_assign', 'refresh\_token', and 'access\_token' parameters are included to authenticate the requests.